G3-real: PASS — distilled edges perfectly cache real 2-hop chains, guard refuses fabrication; G2 dtype fix
Amended-G3 verdict: distilled edge is a perfect chain cache (1.00
agreement, including the unseen paraphrase template); ground-truth
accuracy 0.775/0.786 matches the chain ceiling vs base-rate control
0.42/0.29; fabrication guard holds (0.15 vs 0.42). Write-up in
GATE3_RESULTS.md, raw numbers in real/gate3_results.json. Also a
one-line bfloat16->float fix in gate2_mzkernel.py embed_all.

+++ virtualmesh/gates/real/gate2_mzkernel.py
@@ -53,7 +53,7 @@ def embed_all(model, tok, probes, layers):
         enc = tok(p, return_tensors="pt")
         hs = model(**enc, output_hidden_states=True).hidden_states
         for L in layers:
-            outs[L].append(hs[L][0, -1].numpy())
+            outs[L].append(hs[L][0, -1].float().numpy())
     return {L: np.array(v) for L, v in outs.items()}
